refactor: 모바일 헤더와 데스크탑 헤더로 분리
